bhyve/SMBIOS: use bhyve_config for SMBIOS strings
Some software uses SMBIOS entries to identify the system on which it's
running. In order to make it possible to use such software inside a VM,
SMBIOS entries should be configurable. Therefore, bhyve_config can be
used. While only a few SMBIOS entries might be of interest, it makes
sense that all SMBIOS entries are configurable. This way all SMBIOS
tables are build the same way and there's no special handling for some
tables.

Signed-off-by: Corvin Köhne <c.koehne@beckhoff.com>
Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>