bhyve/bootindex: add bootindex option for several devices
The bootindex option is used to specify the boot order. If no device
uses the bootindex option, the boot order will be determined by
the firmware. If one or more devices specify a bootindex, the
first bootable device with the lowest bootindex will be booted.
It's not garanteed that devices without a bootindex will be
recognized as bootable from the firmware in that case.

Signed-off-by: Corvin Köhne <c.koehne@beckhoff.com>
Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>